feat: moved schedule generation in a separate class
