+++ src/main/java/cybercooker/scheduleservice/grpc/RecipeGateway.java
@@ -5,6 +5,6 @@ import cybercooker.scheduleservice.entity.filter.Filter;
 
 import java.util.List;
 
-public interface RecipeGrpcClientInterface {
+public interface RecipeGateway {
     List<RecipeDTO> getRecipesByFilter(Filter filter, int spaceId);
 }
