+++ src/main/java/cybercooker/scheduleservice/logic/ScheduleGenerator.java
@@ -1,0 +1,8 @@
+package cybercooker.scheduleservice.logic;
+
+import cybercooker.scheduleservice.entity.week.Week;
+
+public interface ScheduleGenerator {
+    Week generateSchedule();
+
+}
